Add file storage resources for Azure
This commit adds the necessary Terraform code to create file storage resources in Azure. It includes the creation of a storage account with a globally unique name and a file share named "fslogix". The storage account is set to use premium tier and locally redundant storage (LRS) replication. The file share has a quota of 5120 GB.

The commit also includes the addition of output variables for the storage account and the file share.